Following beginning the data analysis discovered some errors had been made in the coding process, and realised it's better to treat frequency as the raw number of memories against the total number of memories.
The errors relate to some studies coded whilst tired and map on to an offline record kept of these. All studies are now being rechecked however. Similarly, thought of how it's possible to include dichotmous study scales for perspective use.

The markdown file now reflects some of these alterations and additional calculations. Furthermore, I've improved the formatting so each function is an isolated chunk.

Amendment: errors identified when beginning inspection of the qualtrics export, not the data analysis itself.